refactoring è delete_jobs endpoint
